Add option: whether to reload page on accept/decline/reset click.
I personally find websites that force the page to refresh when you dismiss the cookie banner/message to be rather annoying. I can see that refreshing could be a simple solution for running various scripts to ensure that the cookie settings are applied, but at least for my current project that's entirely unnecessary, so being able to disable that functionality is desirable!